Expand replace regex to include US country code
Expanding the replace regex with alternation to match and trim the US country code variations from US phone numbers. For example, solely the shorthand \D does not match the country code in +1 (555) 555-5555. This new regex commit considers the following ways that the US country code would be added to a 10-digit US phone number:

+1 (555) 555-5555
1 (555) 555-5555
1(555) 555-5555
15555555555